1. Merging all the functionalities from calculate_IV and p_test into create_IV. 2. create_bar_asis and create_IV updated with description and comments.
